o  ID related updates to Continuum facade, Configuration service and notifier Wagon
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497922 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-notifiers/continuum-notifier-wagon/src/main/java/org/apache/maven/continuum/notification/wagon/WagonContinuumNotifier.java
@@ -293,7 +293,7 @@ public class WagonContinuumNotifier
         throws ContinuumException
     {
         File projectWorkingDir =
-            new File( configurationService.getWorkingDirectory(), Integer.toString( project.getId() ) );
+            new File( configurationService.getWorkingDirectory(), Long.toString( project.getId() ) );
         File pomFile = new File( projectWorkingDir, buildDefinition.getBuildFile() );
 
         MavenProject mavenProject = null;
